Revert "code was refactored"
This reverts commit e3818b3a15d3629dddd13f270d94b173c39a9154.

+++ animations/fire.py
@@ -6,7 +6,7 @@ from animations.obstacles import has_collision
 from global_vars import obstacles, obstacles_in_last_collisions
 
 
-async def fire(canvas, border: Border, start_row: float, start_column: float, rows_speed: float = -0.3, columns_speed: float = 0.0):
+async def fire(canvas, border: Border, start_row: int, start_column: int, rows_speed: float = -0.3, columns_speed: float = 0.0):
     """
     Display animation of gun shot, direction and speed can be specified.
